+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+Note: Version 0 of Semantic Versioning is handled differently from version 1 and above. The minor version will be incremented upon a breaking change and the patch version will be incremented for features.
+
 ## [UNRELEASED]
 
 ### Fixed
@@ -18,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- `jupFeeAccount` single wSOL account -> `feeAccounts` map of token mints to token accounts to receive referral fees from both jup and stake pools
 - Use jupiter's `ammsToExclude` param to control which DEXes can't be used
 - Add GooseFX to ammsToExclude list due to too many accounts when trying to swap mSOL -> SOL